// frontend/src/api/generated/models/tag-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * linkweave-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { EntityInfoJson } from './entity-info-json';
import {
    EntityInfoJsonFromJSON,
    EntityInfoJsonFromJSONTyped,
    EntityInfoJsonToJSON,
    EntityInfoJsonToJSONTyped,
} from './entity-info-json';
import type { TagSaveJson } from './tag-save-json';
import {
    TagSaveJsonFromJSON,
    TagSaveJsonFromJSONTyped,
    TagSaveJsonToJSON,
    TagSaveJsonToJSONTyped,
} from './tag-save-json';

/**
 * 
 * @export
 * @interface TagJson
 */
export interface TagJson {
    /**
     * 
     * @type {string}
     * @memberof TagJson
     */
    id: string;
    /**
     * 
     * @type {EntityInfoJson}
     * @memberof TagJson
     */
    entityInfo: EntityInfoJson;
    /**
     * 
     * @type {TagSaveJson}
     * @memberof TagJson
     */
    data: TagSaveJson;
}

/**
 * Check if a given object implements the TagJson interface.
 */
export function instanceOfTagJson(value: object): value is TagJson {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('entityInfo' in value) || value['entityInfo'] === undefined) return false;
    if (!('data' in value) || value['data'] === undefined) return false;
    return true;
}

export function TagJsonFromJSON(json: any): TagJson {
    return TagJsonFromJSONTyped(json, false);
}

export function TagJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): TagJson {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'entityInfo': EntityInfoJsonFromJSON(json['entityInfo']),
        'data': TagSaveJsonFromJSON(json['data']),
    };
}

export function TagJsonToJSON(json: any): TagJson {
    return TagJsonToJSONTyped(json, false);
}

export function TagJsonToJSONTyped(value?: TagJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'entityInfo': EntityInfoJsonToJSON(value['entityInfo']),
        'data': TagSaveJsonToJSON(value['data']),
    };
}