// frontend/src/api/generated/models/tag-save-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * linkweave-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TagSaveJson
 */
export interface TagSaveJson {
    /**
     * 
     * @type {string}
     * @memberof TagSaveJson
     */
    collectionId: string;
    /**
     * 
     * @type {string}
     * @memberof TagSaveJson
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof TagSaveJson
     */
    color?: string;
}

/**
 * Check if a given object implements the TagSaveJson interface.
 */
export function instanceOfTagSaveJson(value: object): value is TagSaveJson {
    if (!('collectionId' in value) || value['collectionId'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function TagSaveJsonFromJSON(json: any): TagSaveJson {
    return TagSaveJsonFromJSONTyped(json, false);
}

export function TagSaveJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): TagSaveJson {
    if (json == null) {
        return json;
    }
    return {
        
        'collectionId': json['collectionId'],
        'name': json['name'],
        'color': json['color'] == null ? undefined : json['color'],
    };
}

export function TagSaveJsonToJSON(json: any): TagSaveJson {
    return TagSaveJsonToJSONTyped(json, false);
}

export function TagSaveJsonToJSONTyped(value?: TagSaveJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'collectionId': value['collectionId'],
        'name': value['name'],
        'color': value['color'],
    };
}